+++ assertions.go
@@ -73,3 +73,27 @@ func AssertPanics(t *testing.T, title string, f func()) {
 
 	f()
 }
+
+// 判断对象(pointer, channel, func, interface, map, slice)是否为 nil
+// nil 是一个 Type 类型的变量, Type 类型是基于 int 的类型
+// var 若变量本身是指针, 占用 8 字节, 指向类型内部结构体并置 0, 仅定义了变量本身, 此时为 nil
+//     指针是非复合类型, 赋值 nil 时, 将 8 字节置 0
+//     map, channel: var 时仅定义了指针, 需要 make 初始化内部结构后才能使用, make 后非 nil
+// var 若变量非指针, 如 struct, int, 非 nil
+// slice:
+//   type slice struct, 占用 24 字节, 1 指针(array unsafe.Pointer) 2 个整型字段(len, cap int)
+//   var 定义后即可使用, 置 0 并分配, 此时 array 指针为 0 即没有实际数据时为 nil
+// interface:
+//   type iface struct(interface 类型), type eface struct(空接口), 占用 16 字节
+//   判断 data 指针为 0 即为 nil, 初始化后即非 0
+func IsNil(i interface{}) bool {
+	if i == nil {
+		return true
+	}
+
+	defer func() {
+		recover()
+	}()
+
+	return reflect.ValueOf(i).IsNil()
+}